Update README.md
I made two branches earlier named readme-edits and readme-edits2. I merged readme-edits with main and it worked fine. Then I had written different content into readme-edts2 and when I merged this with main, I got a conflict message. The only way to resolve the conflict was to erase the entire text from readme-edits and readme-edits two. So now I am creating a new branch called readme-edits3 so that I can merge this with main. I hope this works and I get more used to creating a repository and making commits.